# specs/phase-V-cloud-deployment/contracts/todo-api-openapi.yaml
openapi: 3.0.3
info:
  title: Todo/Chat API
  description: Advanced cloud-native todo chatbot API with event-driven architecture
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.todoapp.com
    description: Production server

paths:
  /api/{user_id}/tasks:
    get:
      summary: Get all tasks for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [all, pending, completed]
          description: Filter tasks by completion status
        - name: priority
          in: query
          required: false
          schema:
            type: string
            enum: [low, medium, high, urgent]
          description: Filter tasks by priority
        - name: tag
          in: query
          required: false
          schema:
            type: string
          description: Filter tasks by tag
        - name: due_after
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Filter tasks with due date after this time
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Number of tasks to return
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Number of tasks to skip
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  total:
                    type: integer
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

    post:
      summary: Create a new task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  /api/{user_id}/tasks/{task_id}:
    get:
      summary: Get a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '404':
          description: Task not found
        '500':
          description: Internal server error

    put:
      summary: Update a task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '404':
          description: Task not found
        '500':
          description: Internal server error

    delete:
      summary: Delete a task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
        '404':
          description: Task not found
        '500':
          description: Internal server error

  /api/{user_id}/tasks/{task_id}/complete:
    post:
      summary: Mark a task as complete
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task marked as complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '404':
          description: Task not found
        '500':
          description: Internal server error

  /api/{user_id}/chat:
    post:
      summary: Send a message to the AI chatbot
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageRequest'
      responses:
        '200':
          description: AI response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

components:
  schemas:
    Task:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique identifier for the task
        user_id:
          type: string
          description: ID of the user who owns the task
        title:
          type: string
          maxLength: 255
          description: Task title
        description:
          type: string
          description: Optional task description
        completed:
          type: boolean
          description: Whether the task is completed
        created_at:
          type: string
          format: date-time
          description: When the task was created
        updated_at:
          type: string
          format: date-time
          description: When the task was last updated
        due_at:
          type: string
          format: date-time
          description: Optional due date/time for the task
        remind_at:
          type: string
          format: date-time
          description: Optional reminder time
        priority:
          type: string
          enum: [low, medium, high, urgent]
          description: Task priority level
        tags:
          type: array
          items:
            type: string
          description: Tags for categorizing the task
        recurrence_pattern:
          type: object
          properties:
            type:
              type: string
              enum: [none, daily, weekly, monthly, yearly]
            interval:
              type: integer
              minimum: 1
            end_condition:
              type: object
              properties:
                after_occurrences:
                  type: integer
                until_date:
                  type: string
                  format: date-time
        parent_task_id:
          type: integer
          description: ID of the parent task if this is a recurring occurrence
        next_occurrence_id:
          type: integer
          description: ID of the next occurrence in the series

    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        due_at:
          type: string
          format: date-time
        remind_at:
          type: string
          format: date-time
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
        tags:
          type: array
          items:
            type: string
        recurrence_pattern:
          $ref: '#/components/schemas/RecurrencePattern'

    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        completed:
          type: boolean
        due_at:
          type: string
          format: date-time
        remind_at:
          type: string
          format: date-time
        priority:
          type: string
          enum: [low, medium, high, urgent]
        tags:
          type: array
          items:
            type: string
        recurrence_pattern:
          $ref: '#/components/schemas/RecurrencePattern'

    RecurrencePattern:
      type: object
      properties:
        type:
          type: string
          enum: [none, daily, weekly, monthly, yearly]
          default: none
        interval:
          type: integer
          minimum: 1
          default: 1
        end_condition:
          type: object
          properties:
            after_occurrences:
              type: integer
              minimum: 1
            until_date:
              type: string
              format: date-time

    ChatMessageRequest:
      type: object
      required:
        - message
      properties:
        conversation_id:
          type: integer
          description: Existing conversation ID (creates new if not provided)
        message:
          type: string
          description: User's message to the AI
        timezone_offset:
          type: integer
          description: Minutes offset from UTC for time-sensitive requests

    ChatMessageResponse:
      type: object
      required:
        - conversation_id
        - response
      properties:
        conversation_id:
          type: integer
          description: The conversation ID
        response:
          type: string
          description: AI assistant's response
        tool_calls:
          type: array
          items:
            type: object
            properties:
              tool_name:
                type: string
              parameters:
                type: object
              result:
                type: object